requirements/prod.txt
dj-database-url==0.4.2
django-configurations==2.0
django==2.0
djangorestframework-jwt==1.11.0
djangorestframework==3.7.7
drf-yasg[validation]
psycopg2==2.7.3.2
